+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.4.17 - 2021-01-??
 - Added a bit extra logging, when using INSTALL=force
-- Labels added: vanilla, taskpaper, calibre, redeye, lucifer, fantastical, launchbar, klokki, notion, lexarrecoverytool, easeusdatarecoverywizard, lastpass, front
+- Labels added: vanilla, taskpaper, calibre, redeye, lucifer, fantastical, launchbar, klokki, notion, lexarrecoverytool, easeusdatarecoverywizard, lastpass, front, obsidian, telegram, wickrme, wickrpro
 
 ## 0.4.16 - 2021-01-08
 - Added variable `INSTALL` that can be set to `force` if software needs to be installed even though latest version is already installed (it will be a reinstall). Usefull if you need to install the universal build of the same software, if the latest was already installed, and the label just improved to handle the universal version. Brave Browser and Google Chrome ready for this.