+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 ## v0.5 - 2020-
 
-## 0.4.17 - 2021-01-??
+## 0.4.17 - 2021-01-19
 - If any blocking process was closed, Installomator will try to open the app again, after the update process is done
 - Added a bit extra logging, when using INSTALL=force
 - Labels with appNewVersion: 1password7
 - Labels changed for Apple Silicon: vlc
-- Labels added: vanilla, taskpaper, calibre, redeye, lucifer, fantastical, launchbar, klokki, notion, lexarrecoverytool, easeusdatarecoverywizard, lastpass, front, telegram, obsidian, wickrme, wickrpro, omnigraffle6, omnioutliner5, omniplan3, omnipresence, omnidisksweeper, exelbanstats, onionshare
+- Labels added: vanilla, taskpaper, calibre, redeye, lucifer, fantastical, launchbar, klokki, notion, lexarrecoverytool, easeusdatarecoverywizard, lastpass, front, telegram, obsidian, wickrme, wickrpro, omnigraffle6, omnioutliner5, omniplan3, omnipresence, omnidisksweeper, exelbanstats, onionshare, keyboardmaestro
 
 ## 0.4.16 - 2021-01-08
 - Added variable `INSTALL` that can be set to `force` if software needs to be installed even though latest version is already installed (it will be a reinstall). Usefull if you need to install the universal build of the same software, if the latest was already installed, and the label just improved to handle the universal version. Brave Browser and Google Chrome ready for this.